requirements.txt
# Core Dependencies
python-dotenv
openai
httpx
pydantic

# Server Dependencies
fastapi
uvicorn

# Browser Automation (optional)
playwright

# Additional Utilities
python-multipart
loguru
sqlalchemy
python-jose>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1